DBMS-PROJECT/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
mysql-connector-python==8.2.0
dj-database-url==2.1.0
